autoscaler/requirements.txt
asyncpg==0.29.0
redis==5.0.4
python-dotenv==1.0.1
docker==7.1.0
httpx==0.27.0
